tweak architecture
